move EditHistory to own file
